requirements.txt
flask==3.0.3
anthropic==0.40.0
python-docx==1.1.2
pdfplumber==0.11.4
pyyaml==6.0.1
gunicorn==23.0.0
